+++ types/v1/serialization_func.go
@@ -8,7 +8,7 @@ import (
 	"unsafe"
 
 	"github.com/grafana/walqueue/types"
-
+	"github.com/prometheus/prometheus/model/exemplar"
 	"github.com/prometheus/prometheus/model/histogram"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/prompb"
@@ -36,7 +36,7 @@ func (s *Serialization) AddPrometheusMetadata(name string, unit string, help str
 	return nil
 }
 
-func (s *Serialization) AddPrometheusMetric(ts int64, value float64, lbls labels.Labels, h *histogram.Histogram, fh *histogram.FloatHistogram, _ map[string]string) error {
+func (s *Serialization) AddPrometheusMetric(ts int64, value float64, lbls labels.Labels, h *histogram.Histogram, fh *histogram.FloatHistogram, _ exemplar.Exemplar, _ map[string]string) error {
 	tss := s.createTimeSeries(ts, value, lbls, h, fh)
 	s.sg.Series = append(s.sg.Series, tss)
 	return nil