+++ types/v2/format_test.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/grafana/walqueue/types"
-
+	"github.com/prometheus/prometheus/model/exemplar"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/prompb"
 	"github.com/stretchr/testify/require"
@@ -26,7 +26,7 @@ func TestDeserializeAndSerialize(t *testing.T) {
 		})
 	}
 	for i := 0; i < 100; i++ {
-		aErr := s.AddPrometheusMetric(time.Now().UnixMilli(), rand.Float64(), lbls, nil, nil, nil)
+		aErr := s.AddPrometheusMetric(time.Now().UnixMilli(), rand.Float64(), lbls, nil, nil, exemplar.Exemplar{}, nil)
 		require.NoError(t, aErr)
 		aErr = s.AddPrometheusMetadata("name", "unit", "help", "gauge")
 		require.NoError(t, aErr)
@@ -81,7 +81,7 @@ func TestExternalLabels(t *testing.T) {
 		})
 	}
 	for i := 0; i < 100; i++ {
-		aErr := s.AddPrometheusMetric(time.Now().UnixMilli(), rand.Float64(), lbls, nil, nil, externalLabels)
+		aErr := s.AddPrometheusMetric(time.Now().UnixMilli(), rand.Float64(), lbls, nil, nil, exemplar.Exemplar{}, externalLabels)
 		require.NoError(t, aErr)
 	}
 	kv := make(map[string]string)